PROPOSALS_TEAM/requirements.txt
pinecone-client>=5.0.0
sentence-transformers>=3.0.1   # if you run local embeddings
pypdf2>=3.0.0
python-docx>=1.1.0
python-frontmatter>=1.0.1
rapidfuzz>=3.9.0
unidecode>=1.3.8
tqdm>=4.66.4
pandas>=2.2.2
reportlab>=4.1.0               # (optional) PDF utilities
